Add first section of landing page
The landing page will be completed from top to bottom. Flex boxes are
used to position each element. Roboto font is used for text. Unsplash is
used for free-to-use placeholder images. This initial design is likely
to change in the final version of the page. Actual header links may be
added as well.