Add third section: quote
Main flex container is set so column. The quote itself is aligned to
center while the quotee is aligned to flex-end. This one took me a while
because I tried to set a max-width of the main flex container to achieve
the right alignment of the quotee. I settled on using padding to get a
similar effect, but this may not be optimal.